@@ -9,37 +9,37 @@
     "test": "mocha --exit"
   },
   "devDependencies": {
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
-    "@testing-library/react": "^12.1.4",
-    "@types/mocha": "^9.1.0",
-    "@types/node": "^17.0.22",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
-    "@typescript-eslint/eslint-plugin": "^5.16.0",
-    "@typescript-eslint/parser": "^5.16.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
+    "@testing-library/react": "^13.3.0",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^16.0.0",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
     "earl-plugin-dom": "nilscox/earl-plugin-dom",
     "earljs": "^0.2.3",
-    "esbuild-loader": "^2.18.0",
-    "eslint": "^8.11.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "esbuild-loader": "^2.19.0",
+    "eslint": "^8.16.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "html-webpack-plugin": "^5.5.0",
     "jsdom": "^19.0.0",
     "jsdom-global": "^3.0.2",
-    "mocha": "^9.2.2",
-    "prettier": "^2.6.0",
+    "mocha": "^10.0.0",
+    "prettier": "^2.6.2",
     "raf": "^3.4.1",
-    "react-refresh": "^0.11.0",
-    "ts-node": "^10.7.0",
-    "type-fest": "^2.12.1",
-    "typescript": "^4.6.2",
-    "webpack": "^5.70.0",
+    "react-refresh": "^0.13.0",
+    "ts-node": "^10.8.0",
+    "type-fest": "^2.13.0",
+    "typescript": "^4.7.2",
+    "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
+    "webpack-dev-server": "^4.9.0"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   }
 }
